update-check-booking

+++ client/src/Components/Pages/Admin/Product/ImportProduct.jsx
@@ -1,13 +1,14 @@
 import { useState, useEffect } from "react";
 import axios from "axios";
-import { Modal, Button, SelectPicker, DatePicker, Input } from "rsuite";
+import { Modal, Button, SelectPicker, DatePicker, Input, Placeholder, Loader } from "rsuite";
 import { format } from "date-fns";
 import { Config } from "../../../../config/connection";
 import { useProduct } from "../../../../config/selectOption";
-//import { Notification, Alert } from '../../../../SweetAlert2'
+import { Notification, Alert } from '../../../../SweetAlert2'
 import Length from "../../../Feature/Length";
 import SearchQuery from "../../../Feature/searchQuery";
 import Pagination from "../../../Feature/Pagination";
+import { AuthenActions } from "../../../../util";
 
 const Importproduct = () => {
   const api = Config.ApiURL;
@@ -17,6 +18,8 @@ const Importproduct = () => {
   const [searchDate, setSearchDate] = useState(null); // for storing the selected date
   const [currentPage, setCurrentPage] = useState(1);
   const [modalType, setModalType] = useState("add"); // Add or edit
+  const [loading, setLoading] = useState(false);
+  const [loadingSave, setLoadingSave]=useState(false)
 
   const [importproductData, setImportProductData] = useState({
     imp_id: null,
@@ -27,17 +30,22 @@ const Importproduct = () => {
   });
 
   const products = useProduct();
+  const actions = AuthenActions();
 
   useEffect(() => {
     fetchgetImpData();
   }, []);
 
   const fetchgetImpData = async () => {
+    setLoading(true)
     try {
       const res = await axios.get(`${api}/imp`);
       setData(res.data);
     } catch (err) {
       console.error("Failed to fetch importproduct data", err);
+    }
+    finally {
+      setLoading(false)
     }
   };
   const resetForm = () => {
@@ -77,7 +85,6 @@ const Importproduct = () => {
   };
 
   const handleChange = (name, value) => {
-  
     setImportProductData({
       ...importproductData,
       [name]: value,
@@ -103,23 +110,31 @@ const Importproduct = () => {
 
   const handleSubmit = async (e) => {
     e.preventDefault();
- 
+    setLoadingSave(true)
     try {
         await axios.post(`${api}/imp/create`, importproductData);
-        alert(`importproduct ${importproductData._id ? "updated" : "added"} successfully!`);
+        Alert.successData(`${importproductData._id ? "ອັບເດດ" : "ບັນທຶກ"} ຂໍ້ມູນສຳເລັດແລ້ວ!`);
         handleClose();
         fetchgetImpData();
     } catch (err) {
       console.error("Failed to submit importproduct data", err);
+      Notification.error('ບັນທຶກຂໍ້ມູນລົ້ມເຫຼວ!');
+    }
+    finally {
+      setLoadingSave(false)
     }
   };
   const handleDeleteClick = async (imp_id) => {
-    try {
-      await axios.delete(`${api}/imp/${imp_id}`);
-      alert("importproduct deleted successfully!");
-      fetchgetImpData();
-    } catch (err) {
-      console.error("Failed to delete importproduct", err);
+    const isConfirmed = await Alert.confirm("ຕ້ອງການລຶບຂໍ້ມູນນີ້ແທ້ບໍ່?");
+    if (isConfirmed) {
+      try {
+        await axios.delete(`${api}/imp/${imp_id}`);
+        Alert.successData("ລຶບຂໍ້ມູນສຳເລັດແລ້ວ!");
+        fetchgetImpData();
+      } catch (err) {
+        console.error("Failed to delete importproduct", err);
+        Notification.error('ລຶບຂໍ້ມູນລົ້ມເຫຼວ');
+      }
     }
   };
   
@@ -142,28 +157,12 @@ const Importproduct = () => {
 
   return (
     <div id="content" className="app-content">
-      <ol className="breadcrumb float-xl-end">
-        <li className="breadcrumb-item">
-          <a href="javascript:;">Home</a>
-        </li>
-        <li className="breadcrumb-item">
-          <a href="javascript:;">Page Options</a>
-        </li>
-        <li className="breadcrumb-item active">importproduct</li>
-      </ol>
-      <h1 className="page-header"><small>header small text goes here...</small>
-      </h1>
-
       <div className="panel panel-inverse">
-        <div className="panel-heading">
-          <h4 className="panel-title">importproduct Panel</h4>
-        </div>
         <div className="panel-body">
           <div className="row mt-2 justify-content-between">
             <div className="d-md-flex justify-content-between align-items-center dt-layout-start col-md-auto me-auto">
               <Length setLength={setLength} />
             </div>
-
             <div className="d-md-flex justify-content-between align-items-center dt-layout-end col-md-auto ms-auto">
               <SearchQuery searchTerm={searchTerm} setSearchTerm={setSearchTerm}/>
               <div className="mb-2 ms-2">
@@ -171,18 +170,16 @@ const Importproduct = () => {
                 onChange={handleDateSearch} placeholder="Select Date"/>
               </div>
               <div className="actions mb-2">
-                <a href="javarscript:;"
-                  className="btn btn-sm btn-success ms-2"
-                  onClick={handleAddClick}
-                >
+              <a href="javarscript:;" className={`btn btn-sm btn-success ms-2 ${!actions.canCreate ? "disabled" : ""}`}
+                  onClick={actions.canCreate ? () => handleAddClick() : (e) => e.preventDefault()}>
                   <i className="fas fa-user-plus"></i>
                 </a>
               </div>
             </div>
           </div>
-
+          <div style={{ overflowX: 'auto', overflowY:'auto' }}>
           <table id="data-table-default"
-            className="table table-striped table-bordered align-middle text-nowrap">
+            className={`table ${!loading && 'table-striped'} table-bordered align-middle text-nowrap`}>
             <thead>
               <tr>
                 <th className="text-nowrap">ລ/ດ</th>
@@ -198,7 +195,14 @@ const Importproduct = () => {
               </tr>
             </thead>
             <tbody>
-              {paginatedData.map((importproduct, index) => (
+            {loading ? (
+                <tr>
+                  <td colSpan={8} className="text-center">
+                  <Placeholder.Grid rows={7} columns={6} active />
+                  <Loader size='lg'  content="ກຳລັງໂຫລດ..." vertical />
+                  </td>
+                </tr>
+              ): paginatedData.length > 0 ? paginatedData.map((importproduct, index) => (
                 <tr key={importproduct.imp_id}>
                   <td width="1%" className="fw-bold">
                     {startIndex + index + 1}
@@ -213,31 +217,33 @@ const Importproduct = () => {
                   <td>
                     <div className="panel-heading">
                       <div className="btn-group my-n1">
-                        <a
-                          href="javascript:;"
-                          className="btn-primary btn-sm dropdown-toggle"
-                          data-bs-toggle="dropdown"
-                        >
-                          <i className="fas fa-ellipsis"></i>
+                        <a href="javascript:;" className="btn-primary btn-sm dropdown-toggle"
+                          data-bs-toggle="dropdown"><i className="fas fa-ellipsis"></i>
                         </a>
                         <div className="dropdown-menu dropdown-menu-end">
-                          <a href="javascript:;" className="dropdown-item"
-                            onClick={() => handleEditClick(importproduct)}><i className="fas fa-pen-to-square"></i>
-                             Edit</a>
-                          <a href="javascript:;" className="dropdown-item"
-                          onClick={() => handleDeleteClick(importproduct.imp_id)}>
-                            <i className="fas fa-trash"></i>
-                             Delete
+                          <a href="javascript:;" className={`dropdown-item ${!actions.canUpdate ? "disabled" : ""}`}
+                            onClick={actions.canUpdate ? () => handleEditClick(importproduct) : (e) => e.preventDefault()}>
+                            <i className="fas fa-pen-to-square fa-fw"></i>
+                            ແກ້ໄຂ
+                          </a>
+                          <a href="javascript:;" className={`dropdown-item ${!actions.canDelete ? "disabled" : ""}`}
+                            onClick={actions.canDelete ? () => handleDeleteClick(importproduct.imp_id) : (e) => e.preventDefault()}>
+                            <i className="fas fa-trash fa-fw"></i>
+                            ລຶບ
                           </a>
                         </div>
                       </div>
                     </div>
                   </td>
                 </tr>
-              ))}
+              )): (
+                <tr className="text-center">
+                  <td colSpan={8} className="text-red">================ ບໍມີຂໍ້ມູນນຳເຂົ້າສິນຄ້າ ===============</td>
+                </tr>
+              )}
             </tbody>
           </table>
-
+          </div>
           <Pagination
             total={filteredData.length}
             length={length}
@@ -284,11 +290,13 @@ const Importproduct = () => {
             </div>
         </Modal.Body>
         <Modal.Footer>
-          <Button type="submit"  appearance="primary">
-            {modalType === "add" ? "Add" : "Update"}
+        <Button type="submit" disabled={loadingSave} appearance="primary">
+          {loadingSave ? (<Loader content="ກຳລັງບັນທຶກ..."/>):
+          <>{modalType === "add" ? "ບັນທຶກ" : "ອັບບເດດ"}</>
+          }
           </Button>
-          <Button onClick={resetForm} appearance="subtle">
-            Cancel
+          <Button onClick={resetForm} color="red" appearance="primary">
+            ຍົກເລີກ
           </Button>
         </Modal.Footer>
         </form>
